refactor: moved redirect page component
