+++ mod.tsx
@@ -6,9 +6,9 @@ import { type Handler, html, serve, UnoCSS } from "./deps.ts";
 import { TopPage } from "./components/pages/TopPage.tsx";
 import { ListPage } from "./components/pages/ListPage.tsx";
 import { LoginPage } from "./components/pages/LoginPage.tsx";
-import { AuthPage } from "./components/pages/AuthPage.tsx";
 import { DashboardPage } from "./components/pages/DashboardPage.tsx";
-import { AddMoviePage } from "./components/pages/movie/Add.tsx";
+import { AuthPage } from "./components/pages/redirect/AuthPage.tsx";
+import { AddMoviePage } from "./components/pages/redirect/movie/Add.tsx";
 import { NotFoundPage } from "./components/pages/error/NotFoundPage.tsx";
 import { cinelogApi } from "./api/mod.ts";
 
@@ -21,8 +21,10 @@ const handler: Handler = (req) => {
   if (pathname === "/") return TopPage(req);
   if (pathname === "/list") return ListPage(req);
   if (pathname === "/login") return LoginPage(req);
-  if (pathname === "/auth") return AuthPage(req);
   if (pathname === "/dashboard") return DashboardPage(req);
+
+  // リダイレクト画面系
+  if (pathname === "/auth") return AuthPage(req);
   if (pathname === "/movie/add") return AddMoviePage(req);
 
   // API
